
\docType{data}
\name{qog.soc.index}
\alias{qog.soc.index}
\title{Data related to Quality of Government variables, Social Policy version (1946-2011).}
\format{A data frame with 1,076 rows and 7 variables}
\description{
  This dataset lists some information on the variables included in the Quality of Government Social Policy dataset:
}
\details{
  \itemize{
    \item \code{variable}
    \item \code{label}
    \item \code{cs.N}: number of observations (cross-sectional dataset)
    \item \code{ts.N}: number of observations (time series dataset)
    \item \code{ts.min}: first year of measurement (time series dataset)
    \item \code{ts.max}: last year of measurement (time series dataset)
    \item \code{ts.T}: year coverage (time series dataset)
    }
}
\seealso{
  \code{\link{qogfind}} to search the index of a QOG dataset
}
\references{
	Svensson, Richard, Stefan Dahlberg, Staffan Kumlin & Bo Rothstein. 
	2012. \emph{The QoG Social Policy Dataset}, version 4Apr12. 
	University of Gothenburg: The Quality of Government Institute, 
	\url{http://www.qog.pol.gu.se}.
}
\keyword{datasets}
