\docType{data}
\name{qog.std.index}
\alias{qog.std.index}
\title{Data related to Quality of Government variables, Standard version (1946-2012).}
\format{A data frame with 763 rows and 9 variables}
\description{
  This dataset lists some information on the variables included in the Quality of Government Standard dataset:
}
\details{
  \itemize{
    \item \code{variable}
    \item \code{label}
    \item \code{ts.min}: first year of measurement (time series dataset)
    \item \code{ts.max}: last year of measurement (time series dataset)
    \item \code{ts.N}: number of observations  (time series dataset)
    \item \code{ts.T}: year coverage (time series dataset)
    \item \code{cs.N}: number of observations (cross-sectional dataset)
    \item \code{cs.min}: first year of measurement (cross-sectional dataset)
    \item \code{cs.max}: last year of measurement (cross-sectional dataset)
    }
}
\seealso{
  \code{\link{qogfind}} to search the index of a QOG dataset
}
\references{
	Teorell, Jan, Nicholas Charron, Stefan Dahlberg, Soren Holmberg, 
	Bo Rothstein, Petrus Sundin & Richard Svensson. 2013. 
	\emph{The Quality of Government Dataset}, version 15May13. 
	University of Gothenburg: The Quality of Government Institute, 
	\url{http://www.qog.pol.gu.se}.
}
\keyword{datasets}
