\docType{data}
\name{qog.index}
\alias{qog.bas.index}
\alias{qog.index}
\alias{qog.soc.index}
\alias{qog.std.index}
\title{Data related to Quality of Government variables, Standard version (1946-2012)}
\format{a data frame}
\description{
  This dataset lists some information on the variables
  included in the Quality of Government Basic, Standard and
  Social Policy datasets:
}
\details{
  \itemize{ \item \code{variable} \item \code{label} \item
  \code{ts.min}: first year of measurement (time series
  dataset) \item \code{ts.max}: last year of measurement
  (time series dataset) \item \code{ts.N}: number of
  observations (time series dataset) \item \code{ts.T}:
  year coverage (time series dataset) \item \code{cs.N}:
  number of observations (cross-sectional dataset) \item
  \code{cs.min}: first year of measurement (cross-sectional
  dataset) \item \code{cs.max}: last year of measurement
  (cross-sectional dataset) }
}
\references{
  Svensson, Richard, Stefan Dahlberg, Staffan Kumlin & Bo
  Rothstein. 2012. \emph{The QoG Social Policy Dataset},
  version 4Apr12. University of Gothenburg: The Quality of
  Government Institute.

  Teorell, Jan, Nicholas Charron, Stefan Dahlberg, Soren
  Holmberg, Bo Rothstein, Petrus Sundin & Richard Svensson.
  2013. \emph{The Quality of Government Dataset}, version
  15May13. University of Gothenburg: The Quality of
  Government Institute.

  Teorell, Jan. Carl Dahlstrom & Stefan Dahlberg. 2011.
  \emph{The QoG Expert Survey Dataset}. University of
  Gothenburg: The Quality of Government Institute.
}
\seealso{
  \code{\link{qogfind}} to search the index of a QOG
  dataset
}
\keyword{datasets}

