
\docType{data}
\name{qog.bas.index}
\alias{qog.bas.index}
\title{Data related to Quality of Government variables, Basic version (1946-2010).}
\format{A data frame with 148 rows and 7 variables}
\description{
  This dataset lists some information on the variables included in the Quality of Government Basic dataset:
}
\details{
  \itemize{
    \item variable
    \item label
    \item cs.N: number of observations (cross-sectional dataset)
    \item ts.N: number of observations (time series dataset)
    \item ts.min: first year of measurement (time series dataset)
    \item ts.max: last year of measurement (time series dataset)
    \item ts.T: year coverage (time series dataset)
    }
}
\seealso{
  \code{\link{qogfind}} to search the index of a QOG dataset
}
\references{
	Teorell, Jan, Marcus Samanni, Soren Holmberg and Bo Rothstein. 
	2012. \emph{The Quality of Government Basic Dataset made from 
		The QoG Standard Dataset version 6Apr11.} 
	University of Gothenburg: The Quality of Government Institute, 
	\url{http://www.qog.pol.gu.se}.
}
\keyword{datasets}
