\name{qoguse}
\alias{qoguse}
\title{qogdata - Import Quality of Government data into R}
\usage{
  qoguse(...)
}
\arguments{
  \item{...}{arguments supplied to the \code{qogdata}
  function.}
}
\description{
  Alias for the \code{qogdata} function, for maximum
  resemblance to Christoph Thewes' \code{QOG} package for
  Stata.
}
\author{
  Francois Briatte \email{f.briatte@ed.ac.uk}
}
\seealso{
  \code{\link{qogdata}}
}

