\name{xtdata-package}
\alias{xtdata-package}
\docType{package}
\title{
Functions to manipulate Quality of Government data and related material
}
\description{
    This package contains a set of functions to download and manipulate 
    Quality of Government (QOG) datasets and codebooks. It uses a specification 
    for panel data that makes it easy to merge QOG datasets with other series, 
    of which a few examples are provided. 
    Please refer to \url{http://www.qog.pol.gu.se} for details on QOG datasets.
}
\details{
\tabular{ll}{
Package: \tab xtdata\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-08-05\cr
License: \tab GPL-3\cr
URL: \tab \url{https://github.com/briatte/xtdata}\cr
}
}
\author{
Francois Briatte \email{f.briatte@ed.ac.uk}

Maintainer: Francois Briatte \email{f.briatte@ed.ac.uk}
}
\references{
  Svensson, Richard, Stefan Dahlberg, Staffan Kumlin & Bo
  Rothstein. 2012. \emph{The QoG Social Policy Dataset},
  version 4Apr12. 
  University of Gothenburg: The Quality of
  Government Institute.

  Teorell, Jan, Nicholas Charron, Stefan Dahlberg, Soren
  Holmberg, Bo Rothstein, Petrus Sundin & Richard Svensson.
  2013. \emph{The Quality of Government Dataset}, version
  15May13. 
  University of Gothenburg: The Quality of
  Government Institute.

  Teorell, Jan. Carl Dahlstrom & Stefan Dahlberg. 2011. 
  \emph{The QoG Expert Survey Dataset}. 
  University of Gothenburg: The Quality of 
  Government Institute.
}
